switch design system to stable-14.0 branch
The owncloud-design-system has been imported to the "web" repo (=
translations happening in the "web" pipeline of the translation sync).
Since web v6.0 (version in the oCIS 2.0.0 GA release) still uses the
dependency from npmjs.com we created a stable branch in the design
system repo and made that the default branch. Translations of the 14.0.x
design system still need to be synced until we archive the repo in a few
months.